refactor: move to sleep data
