+++ src/pages/MySleepDataTab.tsx
@@ -1,8 +1,14 @@
 import { IonContent, IonHeader, IonPage, IonTitle, IonToolbar } from '@ionic/react';
+import { useHistory } from 'react-router-dom';
 import './MySleepDataTab.css';
 import MySleepDataContainer from "../components/MySleepDataContainer";
 
 const MySleepDataTab: React.FC = () => {
+    const history = useHistory();
+    const handleAddDataClick  = () => {
+          history.push("/food/upload")
+    }
+
     return (
         <IonPage>
             <IonHeader>
@@ -16,7 +22,9 @@ const MySleepDataTab: React.FC = () => {
                         <IonTitle size="large">My sleep data</IonTitle>
                     </IonToolbar>
                 </IonHeader>
-                <MySleepDataContainer name="My sleep data page" />
+                <MySleepDataContainer 
+                  name="My sleep data page"           
+                  handleAddDataClick={handleAddDataClick}/>
             </IonContent>
         </IonPage>
     );
